[fix] : 로그인 및 회원가입 에러 수정
